Use the `Settings` structs settingsMap directly
Previously the `Settings` struct was being passed into the
`loaders.LoadJSON` function, and that function was attempting to
unmarshal into it, as opposed to unmarshaling directly into the
underlying map. This change, along with a change proposed to
github.com/quarnster/util/text, allows for the settings to be properly
unmarshaled for later retrieval.