Detect change of file and reload
Detect any change to a file currently being edited and reload it
automatically. The work gone into implementing this will also go towards
watching and reloading of any settings, plugins, as in issue #195.

This does not currently implement any sort of asking of permission to
reload the file if it has been edited and not saved.